semantic_cms/requirements.txt
Django==1.9.2
django-appconf==1.0.1
django-datetime-widget==0.9.3
django-extensions==1.6.1
django-filter==0.12.0
django-imagekit==3.3
django-markupfield==1.4.0
django-mptt==0.8.2
-e git+https://github.com/wilima/django-dag/#egg=django-dag
-e git+https://github.com/alex/django-taggit.git#egg=django-taggit
-e git+https://github.com/wilima/django-wysiwyg-redactor#egg=django-wysiwyg-redactor
-e git+https://github.com/wilima/django-betterforms#egg=django-betterforms
django-templatetag-sugar==1.0
django-uuidfield==0.5.0
django-widget-tweaks==1.4.1
#django-wysiwyg-redactor==0.4.9
djangorestframework==3.3.2
docutils==0.12
Markdown==2.6.5
pilkit==1.1.13
Pillow==3.1.1
psycopg2==2.6.1
pytz==2015.7
six==1.10.0
tagging==0.2.1
wheel==0.29.0
